+++ src/spiceypy/spiceypy.py
@@ -78,7 +78,7 @@ _SPICE_EK_MAXQSEL = 100  # Twice the 50 in gcc-linux-64
 _SPICE_EK_EKRCEX_ROOM_DEFAULT = 100  # Enough?
 
 
-def warn_deprecated_args(**kwargs) -> None:
+def warn_deprecated_args(**kwargs) -> None:  # pragma: no cover
     keys = list(kwargs.keys())
     values = list(kwargs.values())
     if any(values):
@@ -276,8 +276,7 @@ def cell_time(cell_size) -> SpiceCell:
 
 @spice_error_check
 def appndc(
-    item: Union[str, Iterable[str], ndarray, str_],
-    cell: Union[Cell_Char, SpiceCell],
+    item: Union[str, Iterable[str], ndarray, str_], cell: Union[Cell_Char, SpiceCell]
 ) -> None:
     """
     Append an item to a character cell.
@@ -8819,13 +8818,7 @@ def mtxm(m1: ndarray, m2: ndarray) -> ndarray:
 
 
 @spice_error_check
-def mtxmg(
-    m1: ndarray,
-    m2: ndarray,
-    ncol1: OptionalInt = None,
-    nr1r2: OptionalInt = None,
-    ncol2: OptionalInt = None,
-) -> ndarray:
+def mtxmg(m1: ndarray, m2: ndarray) -> ndarray:
     """
     Multiply the transpose of a matrix with
     another matrix, both of arbitrary size.
@@ -8839,7 +8832,6 @@ def mtxmg(
     :param ncol2: Column dimension of m2.
     :return: Transpose of m1 times m2 (O x M).
     """
-    warn_deprecated_args(ncol1=ncol1, nr1r2=nr1r2, ncol2=ncol2)
     ncol1, ncol2 = len(m1[0]), len(m2[0])
     nr1r2 = len(m1)
     m1 = stypes.to_double_matrix(m1)
@@ -8872,9 +8864,7 @@ def mtxv(m1: ndarray, vin: ndarray) -> ndarray:
 
 
 @spice_error_check
-def mtxvg(
-    m1: ndarray, v2: ndarray, ncol1: OptionalInt = None, nr1r2: OptionalInt = None
-) -> ndarray:
+def mtxvg(m1: ndarray, v2: ndarray) -> ndarray:
     """
     Multiply the transpose of a matrix and
     a vector of arbitrary size.
@@ -8883,11 +8873,8 @@ def mtxvg(
 
     :param m1: Left-hand matrix to be multiplied.
     :param v2: Right-hand vector to be multiplied.
-    :param ncol1: Column dimension of m1 and length of vout.
-    :param nr1r2: Row dimension of m1 and length of v2.
     :return: Product vector m1 transpose * v2.
     """
-    warn_deprecated_args(ncol1=ncol1, nr1r2=nr1r2)
     ncol1 = len(m1[0])
     nr1r2 = len(v2)
     m1 = stypes.to_double_matrix(m1)
@@ -8924,9 +8911,6 @@ def mxm(
 def mxmg(
     m1: Union[ndarray, Iterable[Iterable[float]]],
     m2: Union[ndarray, Iterable[Iterable[float]]],
-    nrow1: OptionalInt = None,
-    ncol1: OptionalInt = None,
-    ncol2: OptionalInt = None,
 ) -> ndarray:
     """
     Multiply two double precision matrices of arbitrary size.
@@ -8935,12 +8919,8 @@ def mxmg(
 
     :param m1: nrow1 X ncol1 double precision matrix.
     :param m2: ncol1 X ncol2 double precision matrix.
-    :param nrow1: Row dimension of m1
-    :param ncol1: Column dimension of m1 and row dimension of m2.
-    :param ncol2: Column dimension of m2
     :return: nrow1 X ncol2 double precision matrix.
     """
-    warn_deprecated_args(nrow1=nrow1, ncol1=ncol1, ncol2=ncol2)
     nrow1, ncol1, ncol2 = len(m1), len(m1[0]), len(m2[0])
     m1 = stypes.to_double_matrix(m1)
     m2 = stypes.to_double_matrix(m2)
@@ -8977,9 +8957,6 @@ def mxmt(
 def mxmtg(
     m1: Union[ndarray, Iterable[Iterable[float]]],
     m2: Union[ndarray, Iterable[Iterable[float]]],
-    nrow1: OptionalInt = None,
-    nc1c2: OptionalInt = None,
-    nrow2: OptionalInt = None,
 ) -> ndarray:
     """
     Multiply a matrix and the transpose of a matrix, both of arbitrary size.
@@ -8988,12 +8965,8 @@ def mxmtg(
 
     :param m1: Left-hand matrix to be multiplied.
     :param m2: Right-hand matrix whose transpose is to be multiplied
-    :param nrow1: Row dimension of m1 and row dimension of mout.
-    :param nc1c2: Column dimension of m1 and column dimension of m2.
-    :param nrow2: Row dimension of m2 and column dimension of mout.
     :return: Product matrix.
     """
-    warn_deprecated_args(nrow1=nrow1, nc1c2=nc1c2, nrow2=nrow2)
     nrow1, nc1c2, nrow2 = len(m1), len(m1[0]), len(m2)
     m1 = stypes.to_double_matrix(m1)
     m2 = stypes.to_double_matrix(m2)
@@ -9028,8 +9001,6 @@ def mxv(m1: ndarray, vin: ndarray) -> ndarray:
 def mxvg(
     m1: Union[ndarray, Iterable[Iterable[float]]],
     v2: Union[ndarray, Iterable[Iterable[float]]],
-    nrow1: OptionalInt = None,
-    nc1r2: OptionalInt = None,
 ) -> ndarray:
     """
     Multiply a matrix and a vector of arbitrary size.
@@ -9038,11 +9009,8 @@ def mxvg(
 
     :param m1: Left-hand matrix to be multiplied.
     :param v2: Right-hand vector to be multiplied.
-    :param nrow1: Row dimension of m1 and length of vout.
-    :param nc1r2: Column dimension of m1 and length of v2.
     :return: Product vector m1*v2
     """
-    warn_deprecated_args(nrow1=nrow1, nc1r2=nc1r2)
     nrow1, nc1r2 = len(m1), len(m1[0])
     m1 = stypes.to_double_matrix(m1)
     v2 = stypes.to_double_vector(v2)
@@ -14570,7 +14538,7 @@ def unorm(v1: ndarray) -> Tuple[ndarray, float]:
 
 
 @spice_error_check
-def unormg(v1: ndarray, ndim: OptionalInt = None) -> Tuple[ndarray, float]:
+def unormg(v1: ndarray) -> Tuple[ndarray, float]:
     """
     Normalize a double precision vector of arbitrary dimension and
     return its magnitude.
@@ -14578,10 +14546,8 @@ def unormg(v1: ndarray, ndim: OptionalInt = None) -> Tuple[ndarray, float]:
     https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/unormg_c.html
 
     :param v1: Vector to be normalized.
-    :param ndim: This is the dimension of v1 and vout.
     :return: Unit vector of v1, Magnitude of v1.
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v1)
     v1 = stypes.to_double_vector(v1)
     vout = stypes.empty_double_vector(ndim)
@@ -14632,9 +14598,7 @@ def vadd(
 
 @spice_error_check
 def vaddg(
-    v1: Union[ndarray, Iterable[float]],
-    v2: Union[ndarray, Iterable[float]],
-    ndim: OptionalInt = None,
+    v1: Union[ndarray, Iterable[float]], v2: Union[ndarray, Iterable[float]]
 ) -> ndarray:
     """
     Add two n-dimensional vectors
@@ -14642,10 +14606,8 @@ def vaddg(
 
     :param v1: First vector to be added.
     :param v2: Second vector to be added.
-    :param ndim: Dimension of v1 and v2.
     :return: v1+v2
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v1)
     v1 = stypes.to_double_vector(v1)
     v2 = stypes.to_double_vector(v2)
@@ -14709,7 +14671,7 @@ def vdist(v1: ndarray, v2: ndarray) -> float:
 
 
 @spice_error_check
-def vdistg(v1: ndarray, v2: ndarray, ndim: OptionalInt = None) -> float:
+def vdistg(v1: ndarray, v2: ndarray) -> float:
     """
     Return the distance between two vectors of arbitrary dimension.
 
@@ -14717,10 +14679,8 @@ def vdistg(v1: ndarray, v2: ndarray, ndim: OptionalInt = None) -> float:
 
     :param v1: ndim-dimensional double precision vector.
     :param v2: ndim-dimensional double precision vector.
-    :param ndim: Dimension of v1 and v2.
     :return: the distance between v1 and v2
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v1)
     v1 = stypes.to_double_vector(v1)
     v2 = stypes.to_double_vector(v2)
@@ -14745,7 +14705,7 @@ def vdot(v1: ndarray, v2: ndarray) -> float:
 
 
 @spice_error_check
-def vdotg(v1: ndarray, v2: ndarray, ndim: OptionalInt = None) -> float:
+def vdotg(v1: ndarray, v2: ndarray) -> float:
     """
     Compute the dot product of two double precision vectors of
     arbitrary dimension.
@@ -14754,10 +14714,8 @@ def vdotg(v1: ndarray, v2: ndarray, ndim: OptionalInt = None) -> float:
 
     :param v1: First vector in the dot product.
     :param v2: Second vector in the dot product.
-    :param ndim: Dimension of v1 and v2.
     :return: dot product of v1 and v2.
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v1)
     v1 = stypes.to_double_vector(v1)
     v2 = stypes.to_double_vector(v2)
@@ -14782,7 +14740,7 @@ def vequ(v1: ndarray) -> ndarray:
 
 
 @spice_error_check
-def vequg(v1: ndarray, ndim: OptionalInt = None) -> ndarray:
+def vequg(v1: ndarray) -> ndarray:
     """
     Make one double precision vector of arbitrary dimension equal to another.
 
@@ -14792,7 +14750,6 @@ def vequg(v1: ndarray, ndim: OptionalInt = None) -> ndarray:
     :param ndim: Dimension of vin (and also vout).
     :return: ndim-dimensional double precision vector set equal to vin.
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v1)
     v1 = stypes.to_double_vector(v1)
     vout = stypes.empty_double_vector(ndim)
@@ -14818,17 +14775,15 @@ def vhat(v1: ndarray) -> ndarray:
 
 
 @spice_error_check
-def vhatg(v1: ndarray, ndim: OptionalInt = None) -> ndarray:
+def vhatg(v1: ndarray) -> ndarray:
     """
     Find the unit vector along a double precision vector of arbitrary dimension.
 
     https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/vhatg_c.html
 
     :param v1: Vector to be normalized.
-    :param ndim: Dimension of v1 (and also vout).
     :return: Unit vector v / abs(v).
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v1)
     v1 = stypes.to_double_vector(v1)
     vout = stypes.empty_double_vector(ndim)
@@ -14931,17 +14886,15 @@ def vlcomg(
 
 
 @spice_error_check
-def vminug(vin: ndarray, ndim: OptionalInt = None) -> ndarray:
+def vminug(vin: ndarray) -> ndarray:
     """
     Negate a double precision vector of arbitrary dimension.
 
     https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/vminug_c.html
 
     :param vin: ndim-dimensional double precision vector to be negated.
-    :param ndim: Dimension of vin.
     :return: ndim-dimensional double precision vector equal to -vin.
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(vin)
     vin = stypes.to_double_vector(vin)
     vout = stypes.empty_double_vector(ndim)
@@ -14981,17 +14934,15 @@ def vnorm(v: ndarray) -> float:
 
 
 @spice_error_check
-def vnormg(v: ndarray, ndim: OptionalInt = None) -> float:
+def vnormg(v: ndarray) -> float:
     """
     Compute the magnitude of a double precision vector of arbitrary dimension.
 
     https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/vnormg_c.html
 
     :param v: Vector whose magnitude is to be found.
-    :param ndim: Dimension of v
     :return: magnitude of v calculated in a numerically stable way
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v)
     v = stypes.to_double_vector(v)
     ndim = ctypes.c_int(ndim)
@@ -15117,9 +15068,7 @@ def vrel(
 
 @spice_error_check
 def vrelg(
-    v1: Union[ndarray, Iterable[float]],
-    v2: Union[ndarray, Iterable[float]],
-    ndim: OptionalInt = None,
+    v1: Union[ndarray, Iterable[float]], v2: Union[ndarray, Iterable[float]]
 ) -> float:
     """
     Return the relative difference between two vectors of general dimension.
@@ -15128,10 +15077,8 @@ def vrelg(
 
     :param v1: First vector
     :param v2: Second vector
-    :param ndim: Dimension of v1 and v2.
     :return: the relative difference between v1 and v2.
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v1)
     v1 = stypes.to_double_vector(v1)
     v2 = stypes.to_double_vector(v2)
@@ -15179,7 +15126,7 @@ def vscl(s: float, v1: ndarray) -> ndarray:
 
 
 @spice_error_check
-def vsclg(s: float, v1: ndarray, ndim: OptionalInt = None) -> ndarray:
+def vsclg(s: float, v1: ndarray) -> ndarray:
     """
     Multiply a scalar and a double precision vector of arbitrary dimension.
 
@@ -15187,10 +15134,8 @@ def vsclg(s: float, v1: ndarray, ndim: OptionalInt = None) -> ndarray:
 
     :param s: Scalar to multiply a vector
     :param v1: Vector to be multiplied
-    :param ndim: Dimension of v1
     :return: Product vector, s*v1.
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v1)
     s = ctypes.c_double(s)
     v1 = stypes.to_double_vector(v1)
@@ -15219,7 +15164,7 @@ def vsep(v1: ndarray, v2: ndarray) -> float:
 
 
 @spice_error_check
-def vsepg(v1: ndarray, v2: ndarray, ndim: OptionalInt = None) -> float:
+def vsepg(v1: ndarray, v2: ndarray) -> float:
     """
     Find the separation angle in radians between two double
     precision vectors of arbitrary dimension. This angle is defined
@@ -15229,10 +15174,8 @@ def vsepg(v1: ndarray, v2: ndarray, ndim: OptionalInt = None) -> float:
 
     :param v1: First vector
     :param v2: Second vector
-    :param ndim: The number of elements in v1 and v2.
     :return: separation angle in radians
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v1)
     v1 = stypes.to_double_vector(v1)
     v2 = stypes.to_double_vector(v2)
@@ -15260,7 +15203,7 @@ def vsub(v1: ndarray, v2: ndarray) -> ndarray:
 
 
 @spice_error_check
-def vsubg(v1: ndarray, v2: ndarray, ndim: OptionalInt = None) -> ndarray:
+def vsubg(v1: ndarray, v2: ndarray) -> ndarray:
     """
     Compute the difference between two double precision
     vectors of arbitrary dimension.
@@ -15269,10 +15212,8 @@ def vsubg(v1: ndarray, v2: ndarray, ndim: OptionalInt = None) -> ndarray:
 
     :param v1: First vector (minuend).
     :param v2: Second vector (subtrahend).
-    :param ndim: Dimension of v1, v2, and vout.
     :return: Difference vector, v1 - v2.
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v1)
     v1 = stypes.to_double_vector(v1)
     v2 = stypes.to_double_vector(v2)
@@ -15302,13 +15243,7 @@ def vtmv(v1: ndarray, matrix: ndarray, v2: ndarray) -> float:
 
 
 @spice_error_check
-def vtmvg(
-    v1: ndarray,
-    matrix: ndarray,
-    v2: ndarray,
-    nrow: OptionalInt = None,
-    ncol: OptionalInt = None,
-) -> float:
+def vtmvg(v1: ndarray, matrix: ndarray, v2: ndarray) -> float:
     """
     Multiply the transpose of a n-dimensional
     column vector a nxm matrix,
@@ -15319,11 +15254,8 @@ def vtmvg(
     :param v1: n-dimensional double precision column vector.
     :param matrix: nxm double precision matrix.
     :param v2: m-dimensional double porecision column vector.
-    :param nrow: Number of rows in matrix (number of rows in v1.)
-    :param ncol: Number of columns in matrix (number of rows in v2.)
     :return: the result of (v1**t * matrix * v2 )
     """
-    warn_deprecated_args(nrow=nrow, ncol=ncol)
     nrow, ncol = len(v1), len(v2)
     v1 = stypes.to_double_vector(v1)
     matrix = stypes.to_double_matrix(matrix)
@@ -15366,17 +15298,15 @@ def vzero(v: ndarray) -> bool:
 
 
 @spice_error_check
-def vzerog(v: ndarray, ndim: OptionalInt = None) -> bool:
+def vzerog(v: ndarray) -> bool:
     """
     Indicate whether a general-dimensional vector is the zero vector.
 
     https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/vzerog_c.html
 
     :param v: Vector to be tested
-    :param ndim: Dimension of v
     :return: true if and only if v is the zero vector
     """
-    warn_deprecated_args(ndim=ndim)
     ndim = len(v)
     v = stypes.to_double_vector(v)
     ndim = ctypes.c_int(ndim)
@@ -15878,11 +15808,7 @@ def xpose6(m: Union[ndarray, Iterable[Iterable[float]]]) -> ndarray:
 
 
 @spice_error_check
-def xposeg(
-    matrix: Union[ndarray, Iterable[Iterable[float]]],
-    nrow: OptionalInt = None,
-    ncol: OptionalInt = None,
-) -> ndarray:
+def xposeg(matrix: Union[ndarray, Iterable[Iterable[float]]]) -> ndarray:
     """
     Transpose a matrix of arbitrary size
     in place, the matrix need not be square.
@@ -15890,11 +15816,8 @@ def xposeg(
     https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/xposeg_c.html
 
     :param matrix: Matrix to be transposed
-    :param nrow: Number of rows of input matrix.
-    :param ncol: Number of columns of input matrix
     :return: Transposed matrix
     """
-    warn_deprecated_args(nrow=nrow, ncol=ncol)
     ncol, nrow = len(matrix[0]), len(matrix)
     matrix = stypes.to_double_matrix(matrix)
     mout = stypes.empty_double_matrix(x=nrow, y=ncol)